+++ 01-Interfaces/Conversion.h
@@ -6,14 +6,14 @@
     Obtener el equivalente en grados Celsius
     de una cantidad de grados Farenheit
 */
-float fahrACelcius(int fahr);
+double fahrACelcius(int fahr);
 
 
 /*  
     Obtener el equivalente en grados Farenheit
     de una cantidad de grados Celsius
 */
-float celciusAFahr(int celsius);
+double celciusAFahr(int celsius);
 
 
 #endif
